:wrench: 캐로셀 스타일 코드 분리
